+++ Examples/private_shards.py
@@ -1,0 +1,15 @@
+from awesomeNations import AwesomeNations
+from awesomeNations.awesomeTools import Authentication
+from dotenv import load_dotenv
+from pprint import pp
+import os
+
+# Get data from .env file
+load_dotenv()
+password = os.environ["MY_PASSWORD"]
+
+awesomeAPI = AwesomeNations("My application/1.0.0") # Replace this User-Agent with useful info.
+nation = awesomeAPI.Nation("your nation name here!", Authentication(password))
+
+data = nation.get_shards(('notices', 'ping', 'unread'))
+pp(data)